Rename placeholders.py to prompt.py.
That is its current contents.
