+++ demo_plugin/newhelm/tests/demo_02_unpacking_dependency_test.py
@@ -7,7 +7,7 @@ from newhelm.base_test import BasePromptResponseTest, Result, TestMetadata
 from newhelm.data_packing import TarPacker
 from newhelm.dependency_helper import DependencyHelper
 from newhelm.external_data import ExternalData, WebData
-from newhelm.placeholders import Prompt
+from newhelm.prompt import Prompt
 from newhelm.single_turn_prompt_response import (
     TestItemAnnotations,
     MeasuredTestItem,
