+++ tests/test_records.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 from newhelm.annotation import Annotation
 from newhelm.base_test import Result
-from newhelm.placeholders import Prompt, SUTOptions
+from newhelm.prompt import Prompt, SUTOptions
 from newhelm.records import TestItemRecord, TestRecord
 from newhelm.single_turn_prompt_response import (
     PromptInteraction,
